refactor(git): extract git size-related operations into size_ops module
- Moved measurement size and repository stats functions from size.rs to a new
  git/size_ops.rs module.
- Updated size.rs to import these from git::size_ops, removing duplicated code.
- Added size_ops module to git/mod.rs to expose the functions publicly.

This improves code organization by separating git-related size operations into their
own module, facilitating better modularity and maintainability.
