+++ git_perf/src/git/mod.rs
@@ -2,3 +2,4 @@ mod git_definitions;
 pub mod git_interop;
 mod git_lowlevel;
 mod git_types;
+pub mod size_ops;
